Move remaining Amino messages to module folders
